// src/lib/deprecatedRegistration/types.ts
export enum Environment {
  MARITIME = "MARITIME",
  AVIATION = "AVIATION",
  LAND = "LAND",
}

export enum Purpose {
  PLEASURE = "PLEASURE",
  COMMERCIAL = "COMMERCIAL",
}

export enum Activity {
  // Maritime
  SAILING = "SAILING",
  MOTOR = "MOTOR",
  ROWING = "ROWING",
  SMALL_UNPOWERED = "SMALL_UNPOWERED",
  FISHING_VESSEL = "FISHING_VESSEL",
  MERCHANT_VESSEL = "MERCHANT_VESSEL",
  FLOATING_PLATFORM = "FLOATING_PLATFORM",
  OFFSHORE_WINDFARM = "OFFSHORE_WINDFARM",
  OFFSHORE_RIG_PLATFORM = "OFFSHORE_RIG_PLATFORM",

  // Aviation
  JET_AIRCRAFT = "JET_AIRCRAFT",
  LIGHT_AIRCRAFT = "LIGHT_AIRCRAFT",
  GLIDER = "GLIDER",
  HOT_AIR_BALLOON = "HOT_AIR_BALLOON",
  ROTOR_CRAFT = "ROTOR_CRAFT",
  PASSENGER_PLANE = "PASSENGER_PLANE",
  CARGO_AIRPLANE = "CARGO_AIRPLANE",

  // Land/other activities
  DRIVING = "DRIVING",
  CYCLING = "CYCLING",
  CLIMBING_MOUNTAINEERING = "CLIMBING_MOUNTAINEERING",
  SKIING = "SKIING",
  WALKING_HIKING = "WALKING_HIKING",
  WORKING_REMOTELY = "WORKING_REMOTELY",
  WINDFARM = "WINDFARM",

  OTHER = "OTHER",
}